// src/bitcoin/types.ts
export type Block = {
    hash: string;
    confirmations: number;
    height: number;
    version: number;
    versionHex: string;
    merkleroot: string;
    time: number;
    mediantime: number;
    nonce: number;
    bits: string;
    difficulty: number;
    chainwork: string;
    nTx: number;
    previousblockhash: string;
    nextblockhash?: string;
    strippedsize: number;
    size: number;
    weight: number;
};

export type BlockV0 = string;
export interface BlockV1 extends Block {
    tx: Array<string>;
}
export interface BlockV2 extends Block {
    tx: Array<RawTransactionV1>;
}
export interface BlockV3 extends Block {
    tx: Array<RawTransactionV2>;
}

export type BlockResponse = BlockV0 | BlockV1 | BlockV2 | BlockV3;

/** Transaction and RawTransaction */
export type Transaction = {
    is_active_chain: boolean;
    blockhash: string;
    confirmations: number;
    blocktime: number;
    time: number;
    hex: string;
    txid: string;
    hash: string;
    size: number;
    vsize: number;
    weight: number;
    version: number;
    locktime: number;
};
export type TxIn = {
    coinbase?: string;
    txid?: string;
    vout?: number;
    scriptSig?: {
        asm: string;
        hex: string;
    };
    txinwitness?: string[];
    sequence: number;
};
export interface TxInExt extends TxIn {
    prevout: {
        generated: boolean;
        height: number;
        value: number;
        scriptPubKey: {
            asm: string;
            desc: string;
            hex: number;
            address?: string;
            type: string;
        }
    }
}
export type TxOut = {
    value: number;
    n: number;
    scriptPubKey: {
        asm: string;
        hex: string;
        reqSigs: number;
        type: scriptPubkeyType;
        address?: string;
    };
};
export type RawTransactionV0 = string;
export interface RawTransactionV1 extends Transaction {
    vin: TxIn[];
    vout: TxOut[];
};
export interface RawTransactionV2 extends Transaction {
    fee?: number;
    vin: TxInExt[];
    vout: TxOut[];
};
// ...
export type scriptPubkeyType = string;
// ...
export interface TargetBlockHeight {
    block: BlockV2;
    rpc: BitcoinRpc;
    targetTime?: UnixTimestamp;
}

export type BlockHeight = number;